+++ include/prompt.h
@@ -6,14 +6,15 @@
 /*   By: abarthel <abarthel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/06/20 18:33:04 by abarthel          #+#    #+#             */
-/*   Updated: 2019/09/29 14:57:30 by bprunevi         ###   ########.fr       */
+/*   Updated: 2019/10/20 11:25:39 by bprunevi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef PROMPT_H
 # define PROMPT_H
+#include "libft.h"
 
-int	mkprompt(char **buff);
-int mkprompt_quote(char *input, char **buff);
+size_t	mkprompt(char **buff);
+size_t  mkprompt_quote(char *input, char **buff);
 
 #endif
